Add files via upload
Jupyter notebooks used to generate predictions and figures. Note that to speed up runtime for subsequent sessions, I saved predictions in csv's then reloaded them back into the notebook instead of doing regression all over again. CSVs are provided as well. As of now the GP regression does not produce identical results every time it is run, because of the random initialization procedure. This is something that should be fixed.